A parallelizable version of the Buchberger implementation.
The 'grobnerBasis' function now takes two extra parameters:
(1) A "fine grading" w.r.t. which the polynomials are homogeneous
(2) A function that computes the minimal "fine degrees" of a list
    of degrees.

Under the hood, this is implemented by separating the generation of
S-polynomials ('buchbergerGenStep') from the reduction of the
S-polynomials ('buchbergerRedStep'), with the passing around of a
'Buchberger' data type.